Moved MatchHashesAndScoreQuerySuite and HashingQueryPerformanceSuite out of elastiknn-testing
